chore: move legacy data processing to component
